reconnection fix

+++ app/src/main/java/Sapphire/Auth/Authorizor.java
@@ -28,8 +28,10 @@ public class Authorizor implements IAuthorizor{
     }
 
     public boolean registerDevice(String deviceName, String authToken){
+        if(!checkConnection(conn)){
+            return false;
+        }
         CallableStatement stmt = null;
-
         try{
             stmt = conn.prepareCall("INSERT INTO "+tableName+" (Name,AuthToken) VALUES (?,?);");
             stmt.setString(1, deviceName);
@@ -51,8 +53,10 @@ public class Authorizor implements IAuthorizor{
     }
     
     public boolean removeDeviceAuth(int deviceID){
+        if(!checkConnection(conn)){
+            return false;
+        }
         CallableStatement stmt = null;
-
         try{
             stmt = conn.prepareCall("DELETE FROM "+tableName+" WHERE ID=?;");
             stmt.setInt(1, deviceID);
@@ -67,10 +71,12 @@ public class Authorizor implements IAuthorizor{
     }
 
     public boolean checkDeviceExists(int deviceID){
+        if(!checkConnection(conn)){
+            return false;
+        }
         CallableStatement stmt = null;
-        
         try{
-            stmt = conn.prepareCall("SELECT ID FROM "+tableName+" WHERE ID=?");
+            stmt = conn.prepareCall("SELECT ID FROM "+tableName+" WHERE ID=?;");
             stmt.setInt(1, deviceID);
             stmt.execute();
             if(stmt.getResultSet().next()){
@@ -81,28 +87,34 @@ public class Authorizor implements IAuthorizor{
     }
     
     public int checkAuth(String authToken) throws Exception{
+        if(!checkConnection(conn)){
+            throw new Exception("Unable to connect to database");
+        }
         CallableStatement stmt = null;
         ResultSet result = null;
-        
         try{
-            stmt = conn.prepareCall("SELECT ID FROM "+tableName+" WHERE AuthToken=?");
+            stmt = conn.prepareCall("SELECT ID FROM "+tableName+" WHERE AuthToken=?;");
             stmt.setString(1, authToken);
             stmt.execute();
             result = stmt.getResultSet();
             if(result.next()){
-                int id = Integer.getInteger(result.getString("ID"));
+                int id = Integer.parseInt(result.getString("ID"));
                 return id;
             }
-        }catch(Exception e){}
+        }catch(Exception e){
+            System.out.println("checkAuthError: "+e.getLocalizedMessage());
+        }
         throw new Exception("Id is not authorized");
     }
     
     public String getDeviceName(int deviceID) throws Exception{
+        if(!checkConnection(conn)){
+            throw new Exception("Unable to connect to database");
+        }
         CallableStatement stmt = null;
         ResultSet result = null;
-        
         try{
-            stmt = conn.prepareCall("SELECT Name FROM "+tableName+" WHERE ID=?");
+            stmt = conn.prepareCall("SELECT Name FROM "+tableName+" WHERE ID=?;");
             stmt.setInt(1, deviceID);
             stmt.execute();
             result = stmt.getResultSet();
@@ -114,10 +126,12 @@ public class Authorizor implements IAuthorizor{
     }
     
     public HashMap<Integer,String> showAllDevices() throws Exception{
+        if(!checkConnection(conn)){
+            throw new Exception("Unable to connect to database");
+        }
         HashMap<Integer,String> hm = new HashMap<Integer,String>();
         CallableStatement stmt = null;
         ResultSet result = null;
-        
         try{
             stmt = conn.prepareCall("SELECT ID,Name FROM "+tableName+";");
             stmt.execute();
@@ -136,19 +150,29 @@ public class Authorizor implements IAuthorizor{
         throw new Exception("Unable to find any devices");
     }
 
-    private Connection connect()throws Exception{
+    private Connection connect()throws SQLException, Exception{
         try{
             Class.forName("com.mysql.cj.jdbc.Driver");
         }catch(ClassNotFoundException cnfe){
             throw new Exception("idk what happened here");
         }
         Connection conn = null;
-        try{
-            conn = DriverManager.getConnection(dbURL,dbUsername,dbPassword);
-        }catch(SQLException sqle){
-            throw new Exception(sqle.getMessage());
-        }
-
+        conn = DriverManager.getConnection(dbURL,dbUsername,dbPassword);
         return conn;
+    }
+
+    private boolean checkConnection(Connection conn){
+        try{
+            if(conn.isClosed()){
+                connect();
+            }
+        }catch(SQLException sqle){
+            System.err.println(sqle.getMessage());
+            return false;
+        }catch(Exception e){
+            System.err.println("Unable to reconnect");
+            return false;
+        }
+        return true;
     }
 }
